add sqlutil package with identifier utilities for SQL handling
